category_manager/requirements.txt
# Required dependencies for Category Manager
psycopg2-binary>=2.9.9
